+++ examples/resources/pingdirectory_string_token_claim_validation/import.sh
@@ -1,0 +1,2 @@
+# Importing a String Token Claim Validation requires providing the name of all parent resources in the following format
+terraform import pingdirectory_string_token_claim_validation.myStringTokenClaimValidation "[id-token-validator-name]/[token-claim-validation-name]"