+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * **New Resource:** `pingdirectory_search_reference_criteria ` (#208)
 * **New Resource:** `pingdirectory_server_group` (#209)
 * **New Resource:** `pingdirectory_soft_delete_policy ` (#210)
+* **New Resource:** `pingdirectory_token_claim_validation` (#211)
 
 
 # v0.6.0 May 23 2023